Completed server for website

+++ modules/db/textures.py
@@ -1,8 +1,9 @@
-import sqlite3, uuid, traceback
+import sqlite3, traceback
 from modules.errors import *
 from modules.settings import *
 from modules.db.connection import *
 from modules.db.models import *
+from modules.colors import *
 
 #### Validation functions for textures ###
 def generateTextureID():
@@ -17,8 +18,9 @@ def isTextureIDValid(textureID):
 
 
 def textureIDExists(textureID):
-    isTextureIDValid(textureID)
     try:
+        if not isTextureIDValid(textureID):
+            raise InputException("Invalid texture ID")
         con, cur = connect()
         res = cur.execute("SELECT IDTexture FROM texture WHERE IDTexture = :textureID", {"textureID": textureID})
         numElement = len(res.fetchall())
@@ -30,28 +32,58 @@ def textureIDExists(textureID):
 
 ### Data retrival functions ###
 
+def getTextureInformationFromDatabase(textureDatabaseData):
+    return {
+        'IDTexture': textureDatabaseData[0],
+        'IDModel': textureDatabaseData[1],
+        'isDefault': textureDatabaseData[2],
+        'isColor': textureDatabaseData[3],
+        'colorHex': textureDatabaseData[4],
+        'isImage': textureDatabaseData[5],
+        'extension': textureDatabaseData[6]
+    }
+
+
 def getTexturesList(IDModel):
     try:
+        if not isModelIDValid(IDModel):
+            raise InputException("Invalid model ID")
+        if not modelIDExists(IDModel):
+            raise InputException("Model ID does not exist")
         con, cur = connect()
-        res = cur.execute("SELECT IDTexture, isDefault FROM texture WHERE IDModel = :IDModel", {"IDModel": IDModel})
+        res = cur.execute("SELECT * FROM texture WHERE IDModel = :IDModel", {"IDModel": IDModel})
         textureList = res.fetchall()
         closeConnection(con)
+        for i in range(len(textureList)):
+            textureList[i] = getTextureInformationFromDatabase(textureList[i])
         return textureList
     except sqlite3.Error:
         raise SystemException("Something went wrong during the retrieval of the textures list", traceback.format_exc())
 
 def getTextureInfo(IDTexture):
     try:
+        if not isTextureIDValid(IDTexture):
+            raise InputException("Invalid texture ID")
+        if not textureIDExists(IDTexture):
+            raise InputException("Texture ID does not exist")
         con, cur = connect()
         res = cur.execute("SELECT * FROM texture WHERE IDTexture = :IDTexture", {"IDTexture": IDTexture})
         textureInfo = res.fetchone()
         closeConnection(con)
-        return textureInfo
+        textureInfo = getTextureInformationFromDatabase(textureInfo)
     except sqlite3.Error:
         raise SystemException("Something went wrong during the retrieval of the texture info", traceback.format_exc())
 
 def isTextureFromModel(newDefaultTexture, modelID):
     try:
+        if not isTextureIDValid(newDefaultTexture):
+            raise InputException("Invalid texture ID")
+        if not isModelIDValid(modelID):
+            raise InputException("Invalid model ID")
+        if not textureIDExists(newDefaultTexture):
+            raise InputException("Texture ID does not exist")
+        if not modelIDExists(modelID):
+            raise InputException("Model ID does not exist")
         con, cur = connect()
         res = cur.execute("SELECT IDTexture FROM texture WHERE IDModel = :IDModel AND IDTexture = :IDTexture", {"IDModel": modelID, "IDTexture": newDefaultTexture})
         numElement = len(res.fetchall())
@@ -63,20 +95,19 @@ def isTextureFromModel(newDefaultTexture, modelID):
 ### Creation functions ###
 
 def changeDefaultTexture(modelID, newDefaultTexture):
-    # Validate input
-    if not isModelIDValid(modelID):
-        raise InputException("Model ID is not valid")
-    if not isTextureIDValid(newDefaultTexture):
-        raise InputException("Texture ID is not valid")
-    if not modelIDExists(modelID):
-        raise InputException("Model ID does not exist")
-    if not textureIDExists(newDefaultTexture):
-        raise InputException("Texture ID does not exist")
-    if not isTextureFromModel(newDefaultTexture, modelID):
-        raise InputException("Texture ID does not belong to the model")
-
     # Change default texture
     try:
+        # Validate input
+        if not isModelIDValid(modelID):
+            raise InputException("Model ID is not valid")
+        if not isTextureIDValid(newDefaultTexture):
+            raise InputException("Texture ID is not valid")
+        if not modelIDExists(modelID):
+            raise InputException("Model ID does not exist")
+        if not textureIDExists(newDefaultTexture):
+            raise InputException("Texture ID does not exist")
+        if not isTextureFromModel(newDefaultTexture, modelID):
+            raise InputException("Texture ID does not belong to the model")
         con, cur = connect()
         cur.execute("UPDATE texture SET isDefault = 0 WHERE IDModel = :IDModel", {"IDModel": modelID})
         cur.execute("UPDATE texture SET isDefault = 1 WHERE IDTexture = :IDTexture", {"IDTexture": newDefaultTexture})
@@ -85,16 +116,6 @@ def changeDefaultTexture(modelID, newDefaultTexture):
     except sqlite3.Error:
         raise SystemException("Something went wrong during the change of the default texture", traceback.format_exc())
     saveInfoLog(f"Default texture of the model {modelID} has been changed to {newDefaultTexture}")
-
-def isColorValid(color):
-    if len(color) != 7:
-        return False
-    if color[0] != '#':
-        return False
-    for i in range(1, 7):
-        if color[i] not in '0123456789abcdef':
-            return False
-    return True
 
 def createTexture(textureID, modelID, extension, isNewModel=False, isColor=False, colorHEX=None, isImage=False):
     # Validate input
@@ -116,7 +137,7 @@ def createTexture(textureID, modelID, extension, isNewModel=False, isColor=False
         raise InputException("Color HEX is not defined", colorHEX)
     if isColor and not isinstance(colorHEX, str):
         raise InputException("Color HEX is not a string", colorHEX)
-    if isColor and not isColorValid(colorHEX):
+    if isColor and not isColorValidHEX(colorHEX):
         raise InputException("Color HEX is not valid", colorHEX)
     if isImage and colorHEX is not None:
         raise InputException("Color HEX is defined but isImage is true", colorHEX)
@@ -157,13 +178,13 @@ def createTexture(textureID, modelID, extension, isNewModel=False, isColor=False
 
 ### Delete functions ###
 
-def deleteAllTexturesFromModel(modelID):
-    # Validate input
-    if not isModelIDValid(modelID):
-        raise InputException("Model ID is not valid")
-
+def deleteAllTexturesFromModel(modelID, requireModelExists=False):
     # Delete textures
     try:
+        if not isModelIDValid(modelID):
+            raise InputException("Model ID is not valid")
+        if requireModelExists and not modelIDExists(modelID):
+            raise InputException("Model ID does not exist")
         con, cur = connect()
         cur.execute("DELETE FROM texture WHERE IDModel = :modelID", {"modelID": modelID})
         con.commit()
@@ -173,14 +194,12 @@ def deleteAllTexturesFromModel(modelID):
     saveInfoLog(f"All textures of the model {modelID} have been deleted")
 
 def deleteTexture(textureID):
-    # Validate input
-    if not isTextureIDValid(textureID):
-        raise InputException("Texture ID is not valid")
-    if not textureIDExists(textureID):
-        raise InputException("Texture ID does not exist")
-
     # Delete texture
     try:
+        if not isTextureIDValid(textureID):
+            raise InputException("Texture ID is not valid")
+        if not textureIDExists(textureID):
+            raise InputException("Texture ID does not exist")
         con, cur = connect()
         cur.execute("DELETE FROM texture WHERE IDTexture = :textureID",
                     {"textureID": textureID})
